Rename viewport/react-helpers to viewport/react.
